requirements.txt
pip
future
futures

numpy
# when matplotlib is un-broken, this can be de-specifified.
matplotlib==2.2.3
pandas
scikit-learn
scipy

statsmodels
lmfit
tvb-library
tvb-data
deap
scikit-optimize
